+++ typescript/cli/src/context/types.ts
@@ -56,6 +56,9 @@ export interface CommandContext extends Omit<
   altVmProviders: ChainMap<AltVM.IProvider>;
   supportedProtocols: ProtocolType[];
   skipConfirmation: boolean;
+  warpCoreConfig?: WarpCoreConfig;
+  warpDeployConfig?: WarpRouteDeployConfigMailboxRequired;
+  resolvedWarpRouteId?: string;
   // just for evm chains backward compatibility
   signerAddress?: string;
 }
@@ -69,10 +72,12 @@ export interface WriteCommandContext extends Omit<CommandContext, 'key'> {
 
 export interface WarpDeployCommandContext extends WriteCommandContext {
   warpDeployConfig: WarpRouteDeployConfigMailboxRequired;
+  resolvedWarpRouteId?: string;
 }
 export interface WarpApplyCommandContext extends WriteCommandContext {
   warpDeployConfig: WarpRouteDeployConfigMailboxRequired;
   warpCoreConfig: WarpCoreConfig;
+  resolvedWarpRouteId?: string;
 }
 
 export type CommandModuleWithContext<Args> = CommandModule<