+++ typescript/cli/src/tests/ethereum/warp/apply/warp-apply-rebalancing-config.e2e-test.ts
@@ -115,7 +115,6 @@ describe('hyperlane warp apply owner update tests', async function () {
 
     fixture.writeConfigs();
     await evmWarpCommands.deploy(
-      DEFAULT_EVM_WARP_DEPLOY_PATH,
       HYP_KEY_BY_PROTOCOL.ethereum,
       DEFAULT_EVM_WARP_ID,
     );
@@ -273,7 +272,6 @@ describe('hyperlane warp apply owner update tests', async function () {
 
     writeYamlOrJson(DEFAULT_EVM_WARP_DEPLOY_PATH, warpDeployConfig);
     await evmWarpCommands.deploy(
-      DEFAULT_EVM_WARP_DEPLOY_PATH,
       HYP_KEY_BY_PROTOCOL.ethereum,
       DEFAULT_EVM_WARP_ID,
     );