+++ typescript/rebalancer/src/e2e/composite.e2e-test.ts
@@ -135,7 +135,8 @@ describe('CompositeStrategy E2E', function () {
       deployedAddresses.tokens.anvil1,
       deployer,
     );
-    await token.transfer(userAddress, transferAmount.mul(2));
+    const tx = await token.transfer(userAddress, transferAmount.mul(2));
+    await tx.wait();
 
     const transferResult = await executeWarpTransfer(
       context.multiProvider,